rename basic_geometry to shape
