+++ microgen/shape/polyhedron.py
@@ -15,7 +15,7 @@ import cadquery as cq
 import numpy as np
 import pyvista as pv
 
-from .basic_geometry import Shape
+from .shape import Shape
 
 if TYPE_CHECKING:
     from microgen.shape import KwargsGenerateType, Vector3DType
